% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_flowcode_unmix.R
\name{optimize_flowcode_unmix}
\alias{optimize_flowcode_unmix}
\title{Optimize FlowCode spectral unmixing}
\usage{
optimize_flowcode_unmix(
  raw_data,
  unmixed,
  combined_spectra,
  weights,
  pos_thresholds,
  af_idx,
  af_spectra,
  flowcode_ids,
  has_flowcode,
  combo_fret,
  fret_delta_list,
  fret_delta_norms,
  flowcode_combo_logical,
  flowcode_fluors,
  optimize_fluors,
  variants,
  delta_list,
  delta_norms,
  all_fluor_names,
  af_idx_in_spectra,
  k = 10L,
  weighted = FALSE,
  cell_weighting = FALSE,
  cell_weight_regularize = FALSE,
  nthreads = 1L
)
}
\arguments{
\item{raw_data}{Numeric matrix (n_cells x n_detectors)}

\item{unmixed}{Numeric matrix (n_cells x n_fluors)}

\item{combined_spectra}{Numeric matrix (n_fluors x n_detectors)}

\item{weights}{Numeric vector (n_detectors)}

\item{pos_thresholds}{Numeric vector (n_fluors)}

\item{af_idx}{Integer vector (n_cells), 1-based AF indices}

\item{af_spectra}{Numeric matrix (n_af_variants x n_detectors)}

\item{flowcode_ids}{Integer vector (n_cells), 0 = no flowcode}

\item{has_flowcode}{Integer/logical vector (n_cells)}

\item{combo_fret}{List of FRET spectra matrices}

\item{fret_delta_list}{List of delta matrices}

\item{fret_delta_norms}{List of norm vectors}

\item{flowcode_combo_logical}{Integer matrix (n_combos x n_flowcode_fluors)}

\item{flowcode_fluors}{Character vector}

\item{optimize_fluors}{Character vector}

\item{variants}{List of variant matrices per fluorophore}

\item{delta_list}{List of delta matrices per fluorophore}

\item{delta_norms}{List of delta norms per fluorophore}

\item{all_fluor_names}{Character vector}

\item{af_idx_in_spectra}{Integer, 0-based index of AF in spectra}

\item{k}{Integer, number of variants to test}

\item{weighted}{Logical, use weighted least squares}

\item{cell_weighting}{Logical, use cell-specific weights}

\item{cell_weight_regularize}{Logical, regularize cell weights}

\item{nthreads}{Integer, number of threads}
}
\value{
Numeric matrix (n_cells x n_fluors)
}
\description{
High-performance C++ backend for FlowCode-based spectral optimization.
}
